fix(query1): fixed bug in query1 and deleted some unecessary code

+++ trabalho-pratico/src/catalog/passengersCatalog.c
@@ -24,7 +24,7 @@ gint passenger_flightID_compare_func(gconstpointer a, gconstpointer b){
     const Passenger *passenger1 = (const Passenger*)a;
     const int flightID2 = GPOINTER_TO_INT(b);
 
-    short int flightID1 = get_passenger_flightID(*passenger1);
+    int flightID1 = get_passenger_flightID(*passenger1);
 
     if (flightID1 < flightID2) return -1;
     if (flightID1 > flightID2) return 1;
